requirements.txt
aiohttp==3.5.4
async_timeout==3.0.1
discord.py==1.2.5
lxml==4.4.2
